Move collectFilesFrom to file-system-util.ts
